+++ tweets/urls.py
@@ -5,7 +5,7 @@ from . import views
 
 app_name = "tweets"
 urlpatterns = [
-    path("", views.TweetView.as_view(), name="tweet"),
+    path("", views.TweetCreateView.as_view(), name="tweet"),
     path("<int:pk>/edit/", views.TweetEditView.as_view(), name="tweet_edit"),
     path("<int:pk>/delete/", views.TweetDeleteView.as_view(), name="tweet_delete"),
     path(
