+++ user/static/user/style.css
@@ -184,7 +184,7 @@ ul {
   float:left;
   padding: 10px;
   clear: both;
-  height: 230px;
+  height: 10px;
   width: 70%;
   padding-bottom: 1rem;
 }
