+++ user/templates/user/profile/user_profile.html
@@ -56,19 +56,19 @@
                     </div>
                     <div>
                     {% if user.id == page_user.id %}
-                        <form method="POST" action="{% url 'user:tweet_delete' pk=tweet.pk %}">
+                        <form method="POST" action="{% url 'tweets:tweet_delete' pk=tweet.pk %}">
                             {% csrf_token %}
-                                <a href="{% url 'user:tweet_delete' pk=tweet.pk %}"><button style="all:unset; float:right" onclick="return confirm('Are you sure you want to delete this tweet?')">
+                                <a href="{% url 'tweets:tweet_delete' pk=tweet.pk %}"><button style="all:unset; float:right" onclick="return confirm('Are you sure you want to delete this tweet?')">
                                     <img class="material-icon" src='https://cdn-icons-png.flaticon.com/512/2891/2891491.png'>
                                 </button></a>
                         </form>
-                        <a href="{% url 'user:tweet_edit' pk=tweet.pk %}">
+                        <a href="{% url 'tweets:tweet_edit' pk=tweet.pk %}">
                             <img class="material-icon" style="float:right" src='https://upload.wikimedia.org/wikipedia/commons/thumb/6/64/Edit_icon_%28the_Noun_Project_30184%29.svg/1024px-Edit_icon_%28the_Noun_Project_30184%29.svg.png'>
                         </a>
                     {%  endif %}
                     </div>
                 </div>
-                <a style="text-decoration: none;" href="{% url 'user:tweet_detail' username=tweet.user pk=tweet.pk %}">
+                <a style="text-decoration: none;" href="{% url 'tweets:tweet_detail' username=tweet.user pk=tweet.pk %}">
                     <div class="tweet-body">
                         <p>{{ tweet.body }}</p>
                         {% if tweet.image %}<img src='{{ tweet.image.url }}'>{% endif %}
